+++ packages/test/vitest.config.ts
@@ -8,6 +8,7 @@ export default defineConfig({
       "test/freeze/**/*.test.ts",
       "test/immer/**/*.test.ts",
       "test/sync/**/*.test.ts",
+      "test/plugins/**/*.test.ts",
       "test/types/**/*.test.ts",
       "test/react/**/*.test.tsx",
     ],